Add scripting starter pokemon
Added a "Scripted" option to the starter pokemon section that will use a function in the python script to get a list of starters.

The radio button that selects the scripted starter pokemon option will also add an example function for the selection to the current script if no declaration for it is present yet. It also adds the necessary imports for the function to work.